+++ src/routes/DemographicsUpload.js
@@ -7,8 +7,8 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import React, { Component } from 'react';
 import CSVReader from 'react-csv-reader';
-import './App.css';
-import firebase from './config';
+
+import firebase from '../config';
 
 class DemographicsUpload extends Component {
   state = {
